+++ kutana/backends/vkontakte/extensions.py
@@ -1,16 +1,27 @@
 import json
 import warnings
+from ...helpers import uniq_by, pick
 from ...handler import Handler
 from ...routers import MapRouter
 from ...update import UpdateType
 
 
 class PayloadRouter(MapRouter):
-    __slots__ = ()
+    __slots__ = ("possible_key_sets",)
 
     def __init__(self, priority=7):
         """Base priority is 7"""
         super().__init__(priority=priority)
+        self.possible_key_sets = []
+
+    def _update_key_sets(self, obj):
+        if not isinstance(obj, dict):
+            return
+
+        self.possible_key_sets = uniq_by([
+            *self.possible_key_sets,
+            list(sorted(obj.keys()))
+        ], self._to_hashable)
 
     def _to_hashable(self, obj):
         if isinstance(obj, dict):
@@ -25,18 +36,23 @@ class PayloadRouter(MapRouter):
         backend_identity = ctx.backend.get_identity()
 
         if update.type != UpdateType.MSG or backend_identity != "vkontakte":
-            return ()
+            return
 
         message = update.raw["object"]["message"]
 
         try:
             payload = json.loads(message.get("payload", ""))
         except json.JSONDecodeError:
-            return ()
+            return
 
-        return (self._to_hashable(payload),)
+        if isinstance(payload, dict):
+            for key_set in self.possible_key_sets:
+                yield self._to_hashable(pick(payload, key_set))
+        else:
+            yield self._to_hashable(payload)
 
     def add_handler(self, handler, key):
+        self._update_key_sets(key)
         return super().add_handler(handler, self._to_hashable(key))
 
 
@@ -64,6 +80,7 @@ class ActionMessageRouter(MapRouter):
 
         ctx.action_type = action["type"]
         ctx.action = action
+
         return (action["type"],)
 
 
@@ -87,21 +104,34 @@ class VkontaktePluginExtension:
         """
         Decorator for registering coroutine to be called when
         incoming update is message and payload have specified
-        content.
+        content. Excessive fields in objects are ignored. Use
+        strings and numbers for exact matching.
+
+        Context is automatically populated with following values:
+
+        - payload
 
         See :class:`kutana.plugin.Plugin.on_commands` for details
         about 'priority' and 'router_priority'.
         """
 
         def decorator(func):
+            async def wrapper(update, ctx):
+                ctx.payload = json.loads(
+                    update.raw["object"]["message"].get("payload", "")
+                )
+
+                return await func(update, ctx)
+
             for payload in payloads:
                 self.plugin._add_handler_for_router(
                     PayloadRouter,
-                    handler=Handler(func, priority),
+                    handler=Handler(wrapper, priority),
                     handler_key=payload,
                     router_priority=router_priority,
                 )
-            return func
+
+            return wrapper
 
         return decorator
 
@@ -113,7 +143,12 @@ class VkontaktePluginExtension:
     ):
         """
         Decorator for registering coroutine to be called when
-        incoming update is message with action(only for conversations).
+        incoming update is message with action (only for conversations).
+
+        Context is automatically populated with following values:
+
+        - action_type
+        - action
 
         See :class:`kutana.plugin.Plugin.on_commands` for details
         about 'priority' and 'router_priority'.